feat(workspaces): add unified paginated items service
