+++ packages/server/modules/shared/test/unit/paginatedItems.spec.ts
@@ -1,0 +1,9 @@
+describe('paginatedItems @shared', () => {
+  describe('getPaginatedItemsFactory creates a function, that', () => {
+    it('converts undefined cursors to null')
+    it(
+      'calculates a new cursor from createdAt if there could be more items to be queried'
+    )
+    it('returns items and count from dependencies')
+  })
+})
